Refactor: 디테일 페이지 컴포넌트 분리
