+++ src/components/detail/DetailTab/detailTabStyle.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 const TabSection = styled.div`
-  margin: 100px auto 200px;
+  margin: 10rem auto 20rem;
   grid-column: span 2;
   width: 100%;
 `;
@@ -14,14 +14,18 @@ const TabWrap = styled.div`
 `;
 
 const TabContent = styled.div`
-  margin-top: 50px;
+  margin-top: 5rem;
   width: 100%;
-  height: 500px;
-  line-height: 500px;
+  height: 50rem;
+  line-height: 50rem;
   text-align: center;
   background-color: var(--color-brightGrey);
   color: var(--color-grey);
   font-size: 4rem;
+
+  @media screen and (max-width: 768px) {
+    margin-top: 3rem;
+  }
 `;
 
 export { TabSection, TabWrap, TabContent };
